COMPUTER CONTROLLED TAKEOFF!
We've successfully left the ground via commands sent from a computer.
Next step is setting waypoints and getting the drone to move to them
as well as figuring out how to control the gimbal. Also learned we
can take control with the physical drone controller if things ever
go bad.