Close to full functionality!
Fly.py now contains the script we will most likely use in the final
version of our app. It communicates with the iOS app via our Parse
backend to both get the waypoints the user has selected and the
current location of the iOS device. All that is left is gimbal
control, but we feel we are very close to that as well!